Refactor code and enhance user registration flow
