+++ src/storage.rs
@@ -1,5 +1,6 @@
 pub mod add_user_points;
-pub mod database_storage;
+pub mod save_user;
 pub mod remove_user_points;
 pub mod user;
 pub mod user_exists;
+pub mod get_user;
