+++ src/storage/get_user.rs
@@ -1,9 +1,8 @@
-use crate::storage::save_user::save_to_json;
-use crate::storage::user::User;
 use std::collections::HashMap;
 use std::fs;
+use crate::storage::user::User;
 
-pub fn remove_user_points(user_id: &u64, points: u16) -> Option<()> {
+pub fn get_user(user_id: &u64) -> Option<User> {
     let mut enrollments: HashMap<u64, User> = HashMap::new();
 
     // Load existing data
@@ -12,18 +11,16 @@ pub fn remove_user_points(user_id: &u64, points: u16) -> Option<()> {
         if let Ok(user_enrollments_result) = user_enrollments_result {
             enrollments = user_enrollments_result
         } else {
+            println!("Unable to read user registration file");
             return None;
         }
     }
 
     // edit existing data
     if let Some(user) = enrollments.get(user_id) {
-        let mut user = user.clone();
-        user.points -= points as i64;
-        if save_to_json(&user).is_err() {
-            return None;
-        }
+        return Some(user.clone())
     }
-
-    Some(())
+    
+    println!("User did not exist");
+    None
 }
